requirements.txt
pyserial
systeminfo-python3
screen-brightness-control
keyboard



psutil
pillow
pywin32